backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
boto3==1.36.16
botocore==1.36.16
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.8
gitdb==4.0.11
GitPython==3.1.41
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain-aws==0.2.12
langchain-core==0.3.34
langsmith==0.3.7
numpy==2.2.2
orjson==3.10.15
packaging==24.2
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.11.2
setuptools==69.0.3
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.45.3
tenacity==9.0.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
zstandard==0.23.0
